[FLINK-11463][e2e] Make the kafka resources & kafka streaming test cases as a separate sub-module, also remove the BASH streaming kafka tests.
