+++ flink-end-to-end-tests/flink-end-to-end-tests-common-kafka/src/test/resources/log4j-test.properties
