+++ flink-end-to-end-tests/pom.xml
@@ -77,6 +77,7 @@ under the License.
 		<module>flink-tpch-test</module>
 		<module>flink-streaming-kinesis-test</module>
 		<module>flink-elasticsearch7-test</module>
+		<module>flink-end-to-end-tests-common-kafka</module>
 	</modules>
 
 	<build>
